feat: env init allows certificate import
