+++ internal/pkg/deploy/cloudformation/stack/env_test.go
@@ -40,6 +40,7 @@ func TestEnv_Template(t *testing.T) {
 						PrivateSubnetCIDRs: strings.Split(DefaultPrivateSubnetCIDRs, ","),
 						PublicSubnetCIDRs:  strings.Split(DefaultPublicSubnetCIDRs, ","),
 					},
+					ImportCertARNs: nil,
 				}, gomock.Any()).Return(&template.Content{Buffer: bytes.NewBufferString("mockTemplate")}, nil)
 				e.parser = m
 			},
